First working version
This is the first working version of Group Split. There are a lot of
constants in the code (for things like # of files), but it demonstrates
a working program.

This will automatically create/overwrite files with index sequencing.
All output filenames will share the same number of characters, so they
may be easily sorted.